Second Change

+++ controllers/admin.js
@@ -1,23 +1,21 @@
-const Product = require('../models/product');
-
+const Product = require("../models/product");
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
   const price = req.body.price;
   const description = req.body.description;
-  const product = new Product(null, title,  description, price);
+  const product = new Product(null, title, description, price);
   product.save();
-  res.sendStatus(200)
+  res.sendStatus(200);
 };
 
-exports.getEditProduct = (req, res, next) => {
- 
+exports.getProductById = (req, res, next) => {
   const prodId = req.params.productId;
-  Product.findById(prodId, product => {
+  Product.findById(prodId, (product) => {
     if (!product) {
-      return res.send("Product Not Found")
+      return res.send("Product Not Found");
     }
-    res.send(product)
+    res.send(product);
   });
 };
 
@@ -26,24 +24,34 @@ exports.postEditProduct = (req, res, next) => {
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
   const updatedDesc = req.body.description;
-  const updatedProduct = new Product(
-    prodId,
-    updatedTitle,
-    updatedDesc,
-    updatedPrice
-  );
-  updatedProduct.save();
-  res.status(200).send(updatedProduct)
+  Product.findById(prodId, (product) => {
+    if (!product) {
+      return res.send("Product ID Not Found");
+    }
+    const updatedProduct = new Product(
+      prodId,
+      updatedTitle,
+      updatedDesc,
+      updatedPrice
+    );
+    updatedProduct.save();
+    res.status(200).send(updatedProduct);
+  });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.fetchAll(products => {
-    res.send(products)
+  Product.fetchAll((products) => {
+    res.send(products);
   });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.params.productId;
-  Product.deleteById(prodId);
-  res.sendStatus(200)
+  Product.findById(prodId, (product) => {
+    if (!product) {
+      return res.send("Product Not Found");
+    }
+    Product.deleteById(prodId);
+    res.status(200).send("Success");
+  });
 };
